feat(planteles): added validations for planteles and merged edit and new planteles
